+++ addons/interaction_system/src/enums/EAreaInteractionType.cs
@@ -1,0 +1,7 @@
+namespace InteractionSystem.Enums;
+
+public enum EAreaInteractionType
+{
+	Collision,
+	InputAction
+}