+++ addons/interaction_system/src/enums/EHighlightOn.cs
@@ -1,0 +1,8 @@
+namespace InteractionSystem.Enums;
+
+public enum EHighlightOn
+{
+	Always,
+	Focus,
+	Closest
+}